parallel-chunk-solve-mirrors-per-world-and-baked-colliders
Physics, on 10k unanchored boxes in the bench preset:
  Solve      24.66 ms -> 6.68 ms
  pipeline   30.2 ms  -> 12.9 ms

spatial::ChunkMap is a new sparse-chunk partition - one chunk per proxy by
bounds centre, which is a function rather than a relation and is therefore
splittable across workers. The solver reads its contact colouring off that:
rows whose two *movable* bodies share a chunk form a SolverGroup, groups share
no writable body, and the border rows follow on one thread. Immovable bodies
are read-only during a sweep, which is what stops every contact in a scene
with a floor collapsing into one group.

Four sweeps per handover rather than one, because a dispatch costs about as
much as a sweep does. Drift and sink measured at 1/2/4/8; sink is zero up to
four and 28 mm at eight, so four.

Set-up, the body location search and the impulse write-back are dispatched.
Two passes were measured and left serial: the body gather (1375 us -> 2977 us)
and the narrow phase (4.19 ms wall -> 4.07 ms, but 89.5 ms of worker time).
Store::Get contends rather than shares, and both numbers are in the AGENTS.md.

Mirrors are a per-world budget now. workspace.MaxSurfaces defaults to 50, shows
in the properties panel and the script bindings, and the frame keeps the panes
covering the most screen rather than the first ones in the list. The old
ceiling of sixteen lived in an int8_t on Visual and DrawInstance; widening it
cost nothing on either row.

New collision module at L5 holding convex hulls and triangle soups, because
assets is L8 like physics and two modules at one height have to stay disjoint -
geometry had to enter the stack from below. GJK and EPA sit on top of it in
physics, which is the distance function AGENTS.md listed as absent and which
three things were waiting on: hull colliders, mesh colliders, and a real
time-of-impact sweep.

SweepFastBodies clamps a body moving further than its own thinnest half-extent
one millimetre *into* the surface it would have crossed - not in front of it,
which hangs a bullet against a wall forever because nothing overlaps so the
solver never runs.

ShapeKind gains Hull and Mesh, resolved through a scene::CollisionShapes
resource by name. The studio outlines whatever a part actually collides as,
with a stable colour per face.

determinism, replay, architecture and 337 suites green.

+++ mono.client/src/Scene.cpp
@@ -578,7 +578,7 @@ namespace client {
 		// source store and used entirely outside it - `Universe::Enter` is not
 		// re-entrant, and the far world has to be entered to be read.
 		struct Wanted {
-			int8_t Surface = 0;
+			int16_t Surface = 0;
 			engine::core::Name World;
 		};
 
@@ -645,7 +645,7 @@ namespace client {
 		// before anything is cloned because `AppendPortalClones` walks the store
 		// itself, and a walk started inside another walk is a nesting the ECS
 		// does not owe anybody.
-		std::vector<int8_t> returning;
+		std::vector<int16_t> returning;
 
 		for (const Wanted &entry : wanted) {
 			// **The name resolved against the universe, every frame.** A world
@@ -798,7 +798,7 @@ namespace client {
 				// the same rows the copy above read - so what arrives in this
 				// room is the far half of exactly what the far world drew.
 				if (const auto *list = store.Resource<DrawList>()) {
-					for (const int8_t slot : returning) {
+					for (const int16_t slot : returning) {
 						(void)engine::scene::AppendPortalClones(store, slot, list->Instances, drawn);
 					}
 				}
@@ -941,7 +941,7 @@ namespace client {
 				// is the viewpoint error the pass exists to remove. Skipped here
 				// rather than refused in the renderer so the cost of aiming it is
 				// the only thing wasted.
-				const auto claimed = [&](int8_t slot) {
+				const auto claimed = [&](int16_t slot) {
 					for (const engine::render::PortalView &portal : portals) {
 						if (portal.Index == slot) {
 							return true;
